==> Stage 1 - Feature Creation + Data Cleaning.ipynb <==
import pandas as pd 
import numpy as np 
import string 
from sklearn.metrics.pairwise import cosine_similarity 
from sklearn.feature_extraction.text import CountVectorizer
import nltk
from nltk.corpus import stopwords
stop = stopwords.words("english")
from nltk.stem.porter import *
from nltk.stem.porter import PorterStemmer

initial = pd.read_csv("initial_data.csv")

initial.pattern = initial.pattern.fillna("none")

initial.is_salary = initial.is_salary.fillna(0)

initial.head()

## Clean patterns

# remove whitespace and special characters/numbers
initial['pattern'] = initial['pattern'].str.replace("\s{2,}", " ")
initial['pattern'] = initial['pattern'].str.replace("[^a-zA-Z]", " ")

# remove stopwords
initial['pattern'] = initial['pattern'].apply(lambda x: ' '.join(x for x in x.split() if x not in stop)) 

# stem patterns
stemmer = PorterStemmer()
tokenized_pattern = initial['pattern'].apply(lambda x: x.split())
tokenized_pattern = tokenized_pattern.apply(lambda x: [stemmer.stem(i) for i in x]) # stemming
for i in range(len(tokenized_pattern)):
    tokenized_pattern[i] = ' '.join(tokenized_pattern[i])
initial['pattern'] = tokenized_pattern

def remove_duplicates(pattern):
    pattern = pattern.split()
    ulist = []
    [ulist.append(x) for x in pattern if x not in ulist]
    unique = ' '.join(ulist)
    return unique

## remove repeated words 
pattern_unique = []
for pattern in list(initial.pattern): 
    pattern_unique.append(remove_duplicates(pattern))
initial.pattern = pattern_unique

# remove date and number strings 
initial['pattern'] = initial['pattern'].str.replace("yymmdd| yymmdd|yymmdd ", "")
initial['pattern'] = initial['pattern'].str.replace("nnn| nnn|nnn ", "")


initial[initial.pattern == '']

initial.to_csv("temp.csv")

temp = pd.read_csv("temp.csv")

temp[temp.pattern.isnull()]

temp.pattern = temp.pattern.fillna("none")

temp.pattern.isnull().unique()

## Feature Creation

### Frequency of Near Identical Patterns

clean_patterns = list(temp.pattern)

clean_patterns[:10]

# generate feaure to capture frequency with which similar patterns (suggesting same employer i.e likely salaries) 
# occur for each consumer

# bag of words
vectorizer_pattern = CountVectorizer().fit_transform(clean_patterns)
vectors_pattern = vectorizer_pattern.toarray()

# similarity matrix of words 
pattern_sim = cosine_similarity(vectors_pattern) 
pattern_sim_list = list(pattern_sim)

# for each pattern, count the numbeer of patterns in the dataset tha are at least 80% similar, implying same sender
pattern_sim_num = []
for i in range(24272):
    pattern_sim_num.append(pd.DataFrame(pattern_sim_list[i]>0.8)[0].sum())
temp['pattern_sim_num'] = pattern_sim_num    



temp.pattern_sim_num.head(20)

# we assume that patterns with the same similarity count are near identical
# we see that similar counts occur in clusters by consumer, where one consumer will have many matching pattterns, 
# indicative of their salary payments
# each consumer gets 0-10 salary payments, so we count the number of patterns that are 
# near identical WITHIN each consumer

similar = []
for index, row in temp.iterrows():
    search_for = row['pattern_sim_num']
    temp_pattern = temp[temp['consumer_ref'] == row['consumer_ref']]
    temp_pattern = temp_pattern[temp_pattern['pattern_sim_num'] == search_for]
    val = temp_pattern.shape[0] - 1
    similar.append(val)
temp['pattern_is_similar'] = similar

# boolean variable for whether matching transactions within consumer_ref are between 0-10 
# (more would be too frequent for a salary)

temp['pattern_is_similar_bool'] = ((temp['pattern_is_similar'] > 0) 
                                         &(temp['pattern_is_similar'] < 10)).astype(int)


### Frequency of Similar Transaction Amounts

# consistent amounts suggest regular salaries

# count number of rows within customer that have amounts within 10% of each other
similar = []
for index, row in temp.iterrows():
    amount_low = 0.9 * row['amount']
    amount_high = 1.1 * row['amount']
    temp_count = temp[temp['consumer_ref'] == row['consumer_ref']]
    temp_count = temp_count[temp_count['amount'] >= amount_low]
    temp_count = temp_count[temp_count['amount'] <= amount_high]
    val = temp_count.shape[0] - 1
    similar.append(val)
temp['num_similar'] = similar

# boolean for at least 2 similar amount transactions (at least 2 salaries)
temp['num_similar_bool'] = (temp['num_similar'] >2).astype(int)

### Bins for week of month in which salary is received (most salaries are received at month's end)

temp["postdate"] = pd.to_datetime(temp.postdate)
temp['day'] = temp['postdate'].dt.day
temp['month'] = temp['postdate'].dt.month

temp['week_bin_1'] = (temp['day'] <= 7).astype('int64')
temp['week_bin_2'] = ((temp['day'] > 7) & (temp['day'] <= 14)).astype('int64')
temp['week_bin_3'] = ((temp['day'] > 14) & (temp['day'] <= 21)).astype('int64')
temp['week_bin_4'] = ((temp['day'] > 21) & (temp['day'] <= 31)).astype('int64')

### Large transaction amount (indicative of salary)

temp['large_amount'] = temp.amount > 7000
temp.large_amount = temp.large_amount.astype(int)

### Regex boolean variable capturing presence of "sal"/"wage" words in pattern

temp['wagesal_pattern'] = temp.pattern.str.contains('^sal| wag| sal|^wag|salar|^sal$| sal |^sala| sala').astype(int)


temp.pattern

final_data = temp.iloc[:,4:]

final_data.head()

final_data.to_csv("final_data.csv")
