==> Stage 2 - Comparison of Approaches and Model Selection.ipynb <==
# Model Selection

## A comparison of the NLP only, NLP+Features, and Features only apporaches to identify the best model and approach, followed by hyperparameter tuning to obtain the final model

- RidgeClassifier, LogisticRegression, XGBoost, and SVC were found to yield the highest custom accuracy scores in intial analyses (which compared all 15 models suggest by Pycaret, but had a large runtime)
- Thus we consider these 4 models in our cross-validations here

import numpy as np
import pandas as pd
import fasttext
import random
from sklearn.linear_model import RidgeClassifier
from sklearn.linear_model import LogisticRegression
from sklearn.svm import SVC
from xgboost import XGBClassifier
from sklearn.metrics import balanced_accuracy_score
from sklearn.metrics import recall_score

model_fasttext = fasttext.load_model("transactions_repmodel.bin")

def consumer_accuracy(test_frame, prediction_incorrect): 
    tot_incorrect = test_frame[test_frame[prediction_incorrect] == True].groupby('consumer_ref').sum().count().iloc[0]
    tot = test_frame.groupby('consumer_ref').sum().count().iloc[0]
    accuracy = 1 - tot_incorrect / tot
    return accuracy

final_data = pd.read_csv("final_data.csv")

customer_list = list(final_data.consumer_ref.unique())
random.seed(42)
length = len(customer_list)
cv_random_split_sample = random.sample(customer_list, length)
CV = 5

### *Note that the cv column in the cross validations is only showing the mean value of cv (prints 3 when 5 folds were created)*

### NLP Only Approach

def cross_val_nlp_only(CV, model):
    results = pd.DataFrame(columns = ['cv','train_accuracy_custom', 'test_accuracy_custom', 'balanced_accuracy_train', 
                                  'balanced_accuracy_test', 'recall_train', 'recall_test'])
    for i in range(CV):
        size = int(length/CV)
        lb = size*i
        ub = lb+size
        
        test_cons_list= cv_random_split_sample[lb:ub]
        train_cons_list = set(cv_random_split_sample) - set(test_cons_list)
        train_frame = pd.DataFrame() 
        for customer in train_cons_list: 
            frame = final_data[final_data.consumer_ref == customer] 
            train_frame = pd.concat([train_frame, frame], axis=0)
        test_frame = pd.DataFrame() 
        for customer in test_cons_list: 
            frame = final_data[final_data.consumer_ref == customer] 
            test_frame = pd.concat([test_frame, frame], axis=0)
            
        train_vector = train_frame.pattern.apply(model_fasttext.get_sentence_vector)
    
        X_train = pd.DataFrame.from_records(train_vector.array)
        y_train = train_frame["is_salary"]
        model.fit(X_train, y_train)
        
        test_vector = test_frame.pattern.apply(model_fasttext.get_sentence_vector)
        X_test = pd.DataFrame.from_records(test_vector.array)
        y_test = test_frame["is_salary"]
        
        test_frame["salary_predict"] = model.predict(X_test)
        test_frame['incorrect_predictor'] = test_frame['is_salary'] != test_frame['salary_predict']
        train_frame["salary_predict"] = model.predict(X_train)
        train_frame['incorrect_predictor'] = train_frame['is_salary'] != train_frame['salary_predict']
        
        test_accuracy = consumer_accuracy(test_frame, 'incorrect_predictor')
        train_accuracy = consumer_accuracy(train_frame, 'incorrect_predictor')
        bac_train = balanced_accuracy_score(train_frame.is_salary, train_frame.salary_predict)
        bac_test = balanced_accuracy_score(test_frame.is_salary, test_frame.salary_predict)
        rec_train = recall_score(train_frame.is_salary, train_frame.salary_predict)
        rec_test = recall_score(test_frame.is_salary, test_frame.salary_predict)
        
        results_iter = pd.DataFrame([i + 1,train_accuracy, test_accuracy, bac_train, bac_test, rec_train,
                                     rec_test]).T
        results_iter.columns = results.columns
        results = pd.concat([results, results_iter], axis = 0)
    return pd.DataFrame(results).mean().to_dict()

results_cv_nlp_only = {}
metrics_nlp_only = pd.DataFrame(columns = ['cv','train_accuracy_custom', 'test_accuracy_custom', 'balanced_accuracy_train', 
                                  'balanced_accuracy_test', 'recall_train', 'recall_test'])

def add_model_nlp_only(model):
    results_cv_nlp_only[str(model).split('(')[0]] = cross_val_nlp_only(CV, model)
    return pd.concat([metrics_nlp_only,pd.DataFrame.from_dict(results_cv_nlp_only).T])

ridge = RidgeClassifier()
logistic = LogisticRegression()
xgboost = XGBClassifier()
svc = SVC()

add_model_nlp_only(ridge)

add_model_nlp_only(logistic)

add_model_nlp_only(xgboost)

metrics_nlp_only = add_model_nlp_only(svc)

metrics_nlp_only.sort_values(by = ['test_accuracy_custom'], ascending = False)

### NLP+Features

features = ['wagesal_pattern']  # , 'large_amount', 'week_bin_4','num_similar_bool']
# wagesal_pattern was found to yield highest accuracy, other variables commented out are correlated but lowered
# custom accuracy when included

def cross_val_nlp_features(CV, features, model):
    results = pd.DataFrame(columns = ['cv','train_accuracy_custom', 'test_accuracy_custom', 'balanced_accuracy_train', 
                                  'balanced_accuracy_test', 'recall_train', 'recall_test'])
    for i in range(CV):
        size = int(length/CV)
        lb = size*i
        ub = lb+size
        test_cons_list= cv_random_split_sample[lb:ub]
        train_cons_list = set(cv_random_split_sample) - set(test_cons_list)
        train_frame = pd.DataFrame() 
        for customer in train_cons_list: 
            frame = final_data[final_data.consumer_ref == customer] 
            train_frame = pd.concat([train_frame, frame], axis=0)
        test_frame = pd.DataFrame() 
        for customer in test_cons_list: 
            frame = final_data[final_data.consumer_ref == customer] 
            test_frame = pd.concat([test_frame, frame], axis=0)
        train_vector = train_frame.pattern.apply(model_fasttext.get_sentence_vector)
        count_vect_df = pd.DataFrame.from_records(train_vector.array)
        X_train = pd.concat([train_frame.reset_index()[features],
                             count_vect_df], axis=1)
        y_train = train_frame["is_salary"]
        model.fit(X_train, y_train)
        test_vector = test_frame.pattern.apply(model_fasttext.get_sentence_vector)
        test_count_vect_df = pd.DataFrame.from_records(test_vector.array)
        X_test = pd.concat([test_frame.reset_index()[features],
                            test_count_vect_df], axis=1)
        y_test = test_frame["is_salary"]
        test_frame["salary_predict"] = model.predict(X_test)
        test_frame['incorrect_predictor'] = test_frame['is_salary'] != test_frame['salary_predict']
        train_frame["salary_predict"] = model.predict(X_train)
        train_frame['incorrect_predictor'] = train_frame['is_salary'] != train_frame['salary_predict']
        
        test_accuracy = consumer_accuracy(test_frame, 'incorrect_predictor')
        train_accuracy = consumer_accuracy(train_frame, 'incorrect_predictor')
        bac_train = balanced_accuracy_score(train_frame.is_salary, train_frame.salary_predict)
        bac_test = balanced_accuracy_score(test_frame.is_salary, test_frame.salary_predict)
        rec_train = recall_score(train_frame.is_salary, train_frame.salary_predict)
        rec_test = recall_score(test_frame.is_salary, test_frame.salary_predict)
        
        results_iter = pd.DataFrame([i + 1,train_accuracy, test_accuracy, bac_train, bac_test, rec_train,
                                     rec_test]).T
        results_iter.columns = results.columns
        results = pd.concat([results, results_iter], axis = 0)
    return pd.DataFrame(results).mean().to_dict()

results_cv_nlp_features = {}
metrics_nlp_features = pd.DataFrame(columns = ['cv','train_accuracy_custom', 'test_accuracy_custom', 'balanced_accuracy_train', 
                                  'balanced_accuracy_test', 'recall_train', 'recall_test'])

def add_model_nlp_features(model):
    results_cv_nlp_features[str(model).split('(')[0]] = cross_val_nlp_features(CV, features, model)
    return pd.concat([metrics_nlp_features,pd.DataFrame.from_dict(results_cv_nlp_features).T])

add_model_nlp_features(ridge)

add_model_nlp_features(logistic)

add_model_nlp_features(xgboost)

metrics_nlp_features = add_model_nlp_features(svc)

metrics_nlp_features.sort_values(by = ['test_accuracy_custom'], ascending = False)

### Features Only

features_only = ['wagesal_pattern','large_amount', 'num_similar', 'pattern_is_similar_bool','week_bin_4']

def cross_val_features_only(CV, features, model):
    results = pd.DataFrame(columns = ['cv','train_accuracy_custom', 'test_accuracy_custom', 'balanced_accuracy_train', 
                                  'balanced_accuracy_test', 'recall_train', 'recall_test'])
    for i in range(CV):
        size = int(length/CV)
        lb = size*i
        ub = lb+size
        test_cons_list= cv_random_split_sample[lb:ub]
        train_cons_list = set(cv_random_split_sample) - set(test_cons_list)
        train_frame = pd.DataFrame() 
        for customer in train_cons_list: 
            frame = final_data[final_data.consumer_ref == customer] 
            train_frame = pd.concat([train_frame, frame], axis=0)
        test_frame = pd.DataFrame() 
        for customer in test_cons_list: 
            frame = final_data[final_data.consumer_ref == customer] 
            test_frame = pd.concat([test_frame, frame], axis=0)
        X_train = train_frame[features_only]
        y_train = train_frame["is_salary"]
        model.fit(X_train, y_train)
        X_test = test_frame[features_only]
        y_test = test_frame["is_salary"]
        test_frame["salary_predict"] = model.predict(X_test)
        test_frame['incorrect_predictor'] = test_frame['is_salary'] != test_frame['salary_predict']
        train_frame["salary_predict"] = model.predict(X_train)
        train_frame['incorrect_predictor'] = train_frame['is_salary'] != train_frame['salary_predict']
        
        test_accuracy = consumer_accuracy(test_frame, 'incorrect_predictor')
        train_accuracy = consumer_accuracy(train_frame, 'incorrect_predictor')
        bac_train = balanced_accuracy_score(train_frame.is_salary, train_frame.salary_predict)
        bac_test = balanced_accuracy_score(test_frame.is_salary, test_frame.salary_predict)
        rec_train = recall_score(train_frame.is_salary, train_frame.salary_predict)
        rec_test = recall_score(test_frame.is_salary, test_frame.salary_predict)
        
        results_iter = pd.DataFrame([i + 1,train_accuracy, test_accuracy, bac_train, bac_test, rec_train,
                                     rec_test]).T
        results_iter.columns = results.columns
        results = pd.concat([results, results_iter], axis = 0)
    return pd.DataFrame(results).mean().to_dict()

results_cv_features_only = {}
metrics_features_only = pd.DataFrame(columns = ['cv','train_accuracy_custom', 'test_accuracy_custom', 'balanced_accuracy_train', 
                                  'balanced_accuracy_test', 'recall_train', 'recall_test'])

def add_model_features_only(model):
    results_cv_features_only[str(model).split('(')[0]] = cross_val_features_only(CV, features_only, model)
    return pd.concat([metrics_features_only,pd.DataFrame.from_dict(results_cv_features_only).T])

add_model_features_only(ridge)

add_model_features_only(logistic)

add_model_features_only(xgboost)

metrics_features_only = add_model_features_only(svc)

metrics_features_only.sort_values(by = ['test_accuracy_custom'], ascending = False)

## Looking at all three approaches side by side

metrics_nlp_only.sort_values(by = ['test_accuracy_custom'], ascending = False).head(1)

metrics_nlp_features.sort_values(by = ['test_accuracy_custom'], ascending = False).head(1)

metrics_features_only.sort_values(by = ['test_accuracy_custom'], ascending = False).head(1)

### The SVC model applied to the NLP (fasttext)  + features (wagesal_pattern) approach yields the highest custom accuracy, so this is the model we will proceed with. 

## Hyperparameter Tuning for SVC (nlp + features)

from sklearn.model_selection import RandomizedSearchCV 
from scipy import stats

# Generate train and test frames

random.seed(42)
test_cons_list= random.sample(customer_list, 100)
train_cons_list = set(customer_list) - set(test_cons_list)
train_frame = pd.DataFrame()  
for customer in train_cons_list: 
    frame = final_data[final_data.consumer_ref == customer] 
    train_frame = pd.concat([train_frame, frame], axis=0)
test_frame = pd.DataFrame() 
for customer in test_cons_list: 
    frame = final_data[final_data.consumer_ref == customer] 
    test_frame = pd.concat([test_frame, frame], axis=0)

# Get X_rain, y_train, X_test, and y_test

train_vector = train_frame.pattern.apply(model_fasttext.get_sentence_vector)
X_train = pd.DataFrame.from_records(train_vector.array)
X_train = pd.concat([train_frame['wagesal_pattern'].reset_index(),X_train], axis = 1).drop('index', axis = 1)
y_train = train_frame["is_salary"]
          
test_vector = test_frame.pattern.apply(model_fasttext.get_sentence_vector)
X_test = pd.DataFrame.from_records(test_vector.array)
X_test = pd.concat([test_frame['wagesal_pattern'].reset_index(),X_test], axis = 1).drop('index', axis = 1)
y_test = test_frame["is_salary"]

param_rand = {"C": stats.uniform(2, 10),
             "gamma": stats.uniform(0.1, 1)}

# we can't score on custom accuracy, precision yields the next best score
rand_search = RandomizedSearchCV(svc, param_rand, refit = True, verbose = 3, scoring = 'precision', cv = 3, 
                                 random_state = 42, n_jobs = -1) 

rand_search.fit(X_train, y_train)


print(rand_search.best_score_) 
print(rand_search.best_estimator_) 
print(rand_search.best_params_)

predictions = rand_search.predict(X_test) 

from sklearn.metrics import classification_report, confusion_matrix 
print(classification_report(y_test, predictions)) 

test_frame['prediction'] = predictions
test_frame['prediction_incorrect'] = test_frame['is_salary'] != test_frame['prediction']
consumer_accuracy(test_frame,'prediction_incorrect')

### Custom accuracy after hyperparameter tuning is 78%
